# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DARK-GPT API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SignupInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 1 */
  displayName: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  id: number;
  username: string;
  displayName: string;
  token: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AccessKeyInput {
  key: string;
}

export interface AccessKeyResponse {
  success: boolean;
}

export interface ChatInput {
  messages: ChatMessage[];
  userName?: string;
  groqKey?: string;
}

export interface ChatResponse {
  message: string;
  role: string;
}

export interface ErrorResponse {
  error: string;
}