# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DARK-GPT API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: chat
    description: Chat operations
  - name: auth
    description: Authentication operations
paths:
  /access/verify:
    post:
      operationId: verifyAccess
      tags: [access]
      summary: Verify owner access key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessKeyInput"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessKeyResponse"
        "401":
          description: Wrong key

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message to DARK-GPT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    SignupInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        displayName:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 6
      required: [username, displayName, password]

    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]

    AuthResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        token:
          type: string
      required: [id, username, displayName, token]

    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required: [role, content]

    AccessKeyInput:
      type: object
      properties:
        key:
          type: string
      required: [key]

    AccessKeyResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]

    ChatInput:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        userName:
          type: string
        groqKey:
          type: string
      required: [messages]

    ChatResponse:
      type: object
      properties:
        message:
          type: string
        role:
          type: string
      required: [message, role]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]