# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DARK-GPT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  message: string;
  role: string;
}