# lib/api-zod/src/generated/types/signupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DARK-GPT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignupInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 1 */
  displayName: string;
  /** @minLength 6 */
  password: string;
}